# spec/v1/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AMM \u2014 Agent Memory Manager",
    "description": "REST API for the Agent Memory Manager. Provides durable, structured memory for AI agents.",
    "version": "1.0.0"
  },
...
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
    "/v1/mcp": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "MCP-over-HTTP endpoint",
        "description": "Streamable HTTP endpoint for JSON-RPC MCP requests.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JSON-RPC response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/openapi.json": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "OpenAPI specification",
        "description": "Return the OpenAPI 3.0 specification document.",
        "security": [
          {}
        ],
        "responses": {
          "200": {
            "description": "OpenAPI document",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/swagger/": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Swagger UI",
        "description": "Serve the Swagger UI index page.",
        "security": [
          {}
        ],
        "responses": {
          "200": {
            "description": "Swagger UI HTML",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/v1/status": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Status",
        "description": "Retrieve system status counts and DB path.",
        "security": [
          {}
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/StatusResult"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
...
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "UnsupportedMediaType": {
        "description": "Unsupported media type",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "StatusResult": {
        "type": "object",
        "properties": {
          "db_path": {
            "type": "string"
          },
          "initialized": {
            "type": "boolean"
          },
          "event_count": {
            "type": "integer",
            "format": "int64"
          },
          "memory_count": {
            "type": "integer",
            "format": "int64"
          },
          "summary_count": {
            "type": "integer",
            "format": "int64"
          },
          "episode_count": {
            "type": "integer",
            "format": "int64"
          },
          "entity_count": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  }
}